1. Add formatting model interfaces to 'formatting-model-interfaces.d.ts' 2. Move  powerbi api's to 'visuals-api.d.ts' 3. export both modules from index.d.ts file
